milestone1_Project/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
exceptiongroup==1.3.0
fastapi==0.121.2
flatbuffers==25.9.23
h11==0.16.0
httptools==0.7.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
mpmath==1.3.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==6.33.1
pydantic==2.12.4
pydantic_core==2.41.5
PyMatting==1.1.14
pyreadline3==3.5.4
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.37.0
rembg==2.0.67
requests==2.32.5
rpds-py==0.29.0
scikit-image==0.25.2
scipy==1.15.3
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
